Separate resource and datasource configuration interfaces and helpers
